Move duplicated Serialize / Deserialize functions from api / core into MetadataSerializable
